[Refactor] Switched to using a Wrapper generated by lbstanza-wrappers
This moves all of the extern declarations into an auto-generated wrapper
file. This wrapper file allows us to use this library with both a static
and dynamic library configuration.

Additionally, this allows for importing all functions including
functions with function pointer definitions.

This commit also moves the generated packages for the `tidy-tests`
build to a different directory so that we get better incremental
builds. This reduces test run time by an order of magnitude.